Messages Feature DONE :100: :airplane:
Slightly Buggy - The messages are shown to a user no matter who the user is chatting to. Like if user 1 wants to chat with user 2 but user 2 is on user 3 screen, user 2 can see user 1's messages on user 2's chats, but on refresh it does not remain and user 1 does not know about it, just a display bug only at start. Happens because I am not checking if the sender id is equal to selectedUserId. Will fix it tomo :pray: